+++ src/lib.rs
@@ -23,6 +23,7 @@ pub mod config_watcher;
 pub mod ebpf;
 pub mod monitor;
 pub mod blocklist;
+pub mod client_info;
 pub mod net_monitor;
 pub mod net_util;
 pub mod node_uri;